Revert "Merge pull request #610 from advancedtelematic/refactor/PRO-1463/rename-group-routes"
This reverts commit d390635c70d94e17deb0ca9287f376e942d3e4e2, reversing
changes made to 13f9a4109d8e6854a540451cc5c27d00b4bdb491.
