+++ device-registry/src/main/scala/org/genivi/sota/device_registry/GroupInfoRepository.scala
@@ -74,12 +74,4 @@ object GroupInfoRepository extends SlickJsonHelper {
       .map(_.groupName)
       .update(newGroupName)
       .handleSingleUpdateError(Errors.MissingGroupInfo)
-
-  def groupInfoNamespace(groupId: Uuid)(implicit ec: ExecutionContext): DBIO[Namespace] =
-    groupInfos
-      .filter(r => r.id === groupId)
-      .map(_.namespace)
-      .result
-      .failIfNotSingle(Errors.MissingGroupInfo)
-
 }
