+++ device-registry/src/test/scala/org/genivi/sota/device_registry/ResourceSpec.scala
@@ -41,6 +41,8 @@ trait ResourceSpec extends
   // Route
   lazy implicit val route: Route =
     new Routing(namespaceExtractor, messageBus).route
+
+
 }
 
 trait ResourcePropSpec extends PropSpec with ResourceSpec with PropertyChecks
