+++ docs/swagger/sota-device_registry.yml
@@ -46,7 +46,7 @@ paths:
       responses:
         201:
           description: Created - The UUID of the newly-created device
-
+        
   /devices/{uuid}:
     get:
       description: Get the device object for a single device
@@ -132,7 +132,7 @@ paths:
           description: OK
           schema:
             $ref: '#/definitions/SystemInfo'
-  /device_groups:
+  /devices/group_info:
     get:
       description: Returns a list of all groups
       responses:
@@ -142,7 +142,7 @@ paths:
             type: array
             items:
               $ref: '#/definitions/GroupInfo'
-  /device_groups/{uuid}:
+  /devices/{uuid}/group_info:
     post:
       description: Create a group info entry for a namespace/groupName entry
       parameters:
@@ -205,7 +205,7 @@ paths:
           description: OK
           schema:
             $ref: '#/definitions/GroupInfo'
-  /device_groups/{uuid}/rename:
+  /devices/{uuid}/group_info/rename:
     put:
       description: Rename a group
       parameters:
@@ -222,23 +222,6 @@ paths:
       responses:
         200:
           description: OK
-  /device_groups/{uuid}/devices:
-    get:
-      description: return a list of device uuids in group
-      parameters:
-      - name: uuid
-        description: The uuid of an existing group
-        in: path
-        required: true
-        type: string
-      responses:
-        201:
-          description: Created
-          schema:
-            type: array
-            items:
-              $ref: '#/definitions/uuid'
-
   /device_groups/from_attributes:
     post:
       description: Create a group from two devices' common attributes
